+++ source/rok4_tasks/rok4_tasks/manager_based/locomotion/velocity/mdp/__init__.py
@@ -2,4 +2,6 @@
 
 from isaaclab_tasks.manager_based.locomotion.velocity.mdp import *  # noqa: F401, F403
 
+from .actions import *  # noqa: F401, F403
+from .observations import *  # noqa: F401, F403
 from .rewards import *  # noqa: F401, F403